Update exchange-activesync.md
The "Important" note added at the bottom of the article is a request from Sr. SPM Amir Haque. The PG has declared that only Windows devices version 6.1 and below support password recovery but customers are calling Support based on info from this article that only conveys that password recovery is supported for all mobile devices. Please reword the language of the note and fix the formatting as necessary. Thanks!