Clarifying support stance of storage deduplication
Multiple times a month we're asked if data-deduplication performed by a storage array is supported or not because this document calls out data-dedupe as unsupported. The intent of the article was originally to state Windows Server 2012 de-dupe performed at the OS level is unsupported and made no mention of storage layer dedupe. This update is to clarify we do not support OS level dedupe, but storage layer dedupe is supported yet falls under the third-party storage software solutions support policy.