Update manage-database-copies.md
Loose truncation is a feature helping to avoid situation when disk space is low in case drives with logs are full so we can remove transaction logs even if they were not replicated to passive copies. 
However, this description doesn't mention that Loose Truncation respects backups and just wont work if logs are not in backup yet. 
This can confuse some customers who are trying to rely on that feature and would prefer to victimize backups in low disk space condition. Added remarkes about that into the article.