Clarify that attachment name search is a prefix search, not exact search
The doc suggests that one can search for an exact match of an attachment file name, e.g. "annualreport.ppt". However, in reality, a search with attachment:"annualreport.ppt" will return both "annualreport.ppt" and "2017 annualreport.ppt". In fact, as of today, there is apparently no way to perform an exact search for an attachment file name. 
The situation is similar when searching for Subject, but there, the doc correctly mentions that searching for subject:"Quarterly Financials" will return messages that contain the phrase "Quarterly Financials" _anywhere_ in the text of the subject line.