Added ECP to CBA configuration
Experienced some issues when CBA was enabled for OWA but not for ECP on Exchange 2016. We should check if CBA should be enabled for ECP if it was enabled for OWA. 
Get redirect to 
https://<servername>/owa/auth/logon.aspx?replaceCurrent=1&url=https%3a%2f%2fservername%2fecp%2f
for doing auth via 
https://<servername>/owa/auth.owa
which requires CBA because of OWA CBA configuration. Not configuring CBA for ECP will break access to Exchange Admin Center.
We also display the following warning when Set-OWAVirtualDirectory is called to change authentication methods: 

WARNING: You've changed the authentication method for the OWA virtual directory. Please make the corresponding
configuration for the ECP virtual directory in the same web site.

So please clarify internally and merge this pull request if the information above is correct.
If more information is needed please drop me an e-mail to my work account.
Thanks,
/Lukas