Organization Limit is applicable to external sender and recipient as well
Many customers does not know that the limit is application for External Sender and recipient as well. Customers just increase the Max Receive size thinking its enough to receive a higher message size email from Internet. But it won't work as he should increase for Max send size as well because this limit would be applicable for the external Sender. It comes as a surprise for customer when informed, we also got feedback from customer that this is not documented anywhere.