Update mail routing domains verbiage
Currently it is confusing because the document mixes up parts of the
G Suite routing domain instructions with the Office 365 routing domain
instructions. Specifically the "to Office 365" routing domain creation
instructions are mentioned in the "to G Suite" section, and vice-versa.
Also, some sections don't quite match or could be slightly improved.