G Suite Migration provisioning instruction update
Improve the G Suite Migration topic by elaborating on the recommended primary domain to use for the mail users that should be provisioned to correspond to the source G Suite users.  This should hopefully make it clearer that we recommend having the primary email address be at the primary domain.