Caution related to IIS 6.0 support lifecycle
Recently we came across multiple cases where customer is using IIS SMTP to relay email to Office 365 and faced mailflow issues. 
Normally things work fine but there are cases where it becomes little complex. For example, I heard two cases where IIS is hosted in Azure and SMTP is not using the Certificate Associated in the virtual directory and instead using Azure CPR certificate. There is another case that I am working correctly where one random email will stuck in queue and will stay there for ever unless we reset the network connector (its not a network issue as Netmon does not show anything related to that connection at all)

Now as we know the support for IIS 6.0 is long gone with the support of Windows 2003 (https://support.microsoft.com/en-in/lifecycle/search/810) which is July 2015, I am proposing we add a caution in thus article stating the support Lifecycle has ended for IIS 6.0 SMTP component (This is already discussed with Scott Landry and Sean Stevenson)