Wrong information about when both Send As and Send on Behalf permission is enabled
The article have wrong information about when both Send As and Send on Behalf permission is enabled. It says "If a user has both Send As and Send on Behalf permissions to a mailbox or group, the Send on Behalf permission is always used." Actually it's "Send As" that is always used when both permissions are enabled. I removed the line from the Article. Also the behavior should be added as a note below all the permissions are discussed.