Update how-to-set-up-a-multifunction-device-or-application-to-send-email-using-office-3.md
TLS is not must for Direct send. Infact it's mentioned in this same article under "Features of direct send" 

Uses Office 365 to send emails, but does not require a dedicated Office 365 mailbox.
Doesn't require your device or application to have a static IP address. However, this is recommended if possible.
Doesn't work with a connector; never configure a device to use a connector with direct send, this can cause problems.
Doesn't require your device to support TLS.