Update manage-mail-flow-using-third-party-cloud.md
Removed obsolete (and incorrect) reference to IP addresses which do not appear in this article. The 10.x IP address is also a bad example as it is an private IP address (non-public) and should never be configured in such configuration even if it is a fictitious example, it is a bad one as customers servers have both public and private IP addresses and this would make them configure the private one of the two.