Update incorrect info on custom mailtip
In the section for Custom mailtips where it says
By default, custom MailTips aren't displayed if the sender isn't allowed to send messages to the recipient (the Restricted Recipient MailTip is displayed instead). However, you can change this configuration and have the custom MailTip also display.

This is wrongly documented. There's currently no existing config to allow custom mailtip to show on restricted recipient. I also checked with the original owner and here's what he said "When we were designing this, we wanted to avoid overloading the UI with too many messages. And we figured that if you can't send a message to a recipient, then there is no reason to show the custom tip."