Update the table to include 432 4.3.2
We often get support tickets and escalations about this. It's not easy to find info about it on the web. Let's make it easier. 

BTW, we should update the title of this document since it's NOT just about NDRs - none of the 4.x.x errors result in an NDR, or if they do they become 5.x.x errors. Therefore, should say something like  "Non-delivery report and SMTP errors in Exchange Online".