SIR does not retain COW items instead Litigation Hold.
Here is the statement from the article:
Also, if a message is changed by a user or a process, copies of the original item are also retained when single item recovery is enabled. 

Here is what we need to update:
Also, if a message is changed by a user or a process, copies of the original item are also retained when single item recovery and Litigation Hold both are enabled.

This is applicable for both Exchange Server 2013, 2016, 2019 and Exchange Online. All reference articles for Single Item Recovery needs to be updated for related products.