Single item recovery does not preserve item change
Litigation Hold on only enabled the ability to maintain the original item when changed in the user mailbox in the Recoverable Items. Single Item Recovery has nothing to do with maintaining the original item.

Following statement:
Also, if a message is changed by a user or a process, copies of the original item are also retained when single item recovery and Litigation Hold both are enabled.

The current statement implies that Single Item Recovery and Litigation both needs to be enabled to retain original items when it is changed and looks incorrect. So removing these lines will best option instead of making a comment under Enable and Disable Single Item recovery article.