Update journaling.md
Per VSO 1747880, the Exchange PG provided known scenarios in which duplicate journal reports may occur in a hybrid Exchange environment (on-premises and cloud). They requested I create a pull request to include those scenarios in this document. I initially located this information in the Troubleshooting section, but feel free to move it elsewhere and flesh it out further.  I've never hear the term chipping that is noted in scenario 1a, so this may need to be changed to a more common term or further explained in the document. Also, should "duplicate" or "multiple" be used here? In most cases, it would just be duplicate journal reports being generated, one in on-prem and the other in the cloud.