Minor changes done
Question: It is mentioned that there are two sections to this report where as I see 3 points below that statement. So not sure how to read this.
Do you mean to say that the charts make one section and the table makes the other section? If yes, then I think we need to combine point 1 and 2, and point 3 becomes point 2 then.