Comments as below
Feedback

1. Please add the link to the new EAC or the old EAC as the case may be in the relevant procedure. This is to keep
it consistent with other topics where we have provided the link.
2. The sections mentioned after bullet point 3 under "Use the new EAC to change mail-enabled security group properties", 
need to be aligned appropriately under the bulleted list. They now appear to be standalone subtopics which is not the case.
3. "Email options in new EAC" - I believe this should just be "Email options". or add "in new EAC" to other sections as well
again for the consistency. Since all these sections are falling under the new EAC, there is no necessity
to call out a particular section. If this section is specifically added for new EAC, then call it out in the paragraph.