Minor changes
[!IMPORTANT] Mail flow rules are now available in the new Exchange Admin Center.

The above banner does not seem to be adequate. The link goes to the EAC page. We also need to provide a for more information
link to the documentation too.

we recommend using Option 2 or 3 below. - Instead of putting "below", can your put a crossreference to these topics. These topics are further down and require some amount of scrolling before the user can find them. It would be useful to just provide a crossreference here.