Remove references and links to Exchange Server
There are multiple links to Exchange Server DLP content about built in sensitive information types that is out of date as of December 2020. There is a separate article for Exchange Server (https://docs.microsoft.com/en-us/exchange/policy-and-compliance/data-loss-prevention/data-loss-prevention?view=exchserver-2019) and this article exists in Exchange Online area so we should remove references and links to Exchange Server.