Update retention-tags-and-policies.md
Add this From a retention perspective, the Primary Mailbox and Online Archive should not be looked at as separate entities.  Retention settings are applied to the Primary Mailbox, by design, extend to the Online Archive. The Online Archive is an Extension of the Primary Mailbox.  
Remove this:
Archive mailboxes don't have a separate retention policy. The same retention policy is applied to the primary and archive mailbox.