Update disable-basic-authentication-in-exchange-online.md
Clarify that the option to disable within the M365 Admin portal does not disable all the legacy protocol options, and that it also does not apply the configuration to the existing mailboxes. I'm sure this needs technical approval and wordsmithing.  Or maybe it might just generate a technical improvement within the Exch Online functionality.