Update build-numbers-and-release-dates.md
I have seen some confusion out there related to the fact that in build lists, the SUs are listed as 'indented' and some customers thought that they were 'under' the respective CUs. We should somehow indicate that the list is chronological - newest releases will simply always be on the top of the list.