Update manage-dynamic-distribution-groups.md
Admins can be confused and misunderstand "Users with Exchange mailboxes" and "Resource mailboxes" filter. 
They may think the "Users with Exchange mailboxes" only contains accounts with RecipientTypeDetails as UserMailbox, and only the "Resource mailboxes" option filters resource mailboxes.
However the fact is "Users with Exchange mailboxes" option also contains resources mailboxes.
It is better that this potential misunderstanding can be mentioned in document.