Update antimalware-procedures.md
Feel free to propose any changes to the description. At first view from the available anti Malware articles it seems it is not possible to bypass anti-Malware detection for a user subgroup without disabling the detection for all recipients. Maybe the bypass part of this article can have this alternative method described and referenced in the part I edited. Or I can create a new article for this particular scenario.