[BrokenLinksH2] fixing broken link
**Global effort to fix broken links**

@AshaIyengar21

The Content & Learning team is fixing broken links on docs.microsoft.com for the rest of H2. 

This broken link was found. I searched for a replacement, is this a good choice?

This effort will eliminate potential accessibility, security, and usability issues. This PR includes only link fixes and does not change other content. Please review within five business days and merge, or comment in the PR with any changes you'd like to see. Thanks!