[BrokenLinksH2] fixing broken links
**Global effort to fix broken links**

@mattpennathe3rd

The Content & Learning team is fixing broken links on docs.microsoft.com for the rest of H2. This effort will eliminate potential accessibility, security, and usability issues. This PR includes only link fixes and does not change other content. Please review within five business days and merge, or comment in the PR with any changes you'd like to see. Thanks!

I've fixed two broken links in this TOC.  There's a third link that goes to a nested TOC:

  - name: Exchange Server 2013 Technical Articles
    href: /Exchange/Exchange/ExchangeServer2013/exchange-server-2013-technical-articles 

needs to point to this TOC: Exchange/ExchangeServer2013/technical-articles/toc.yml

I'm not sure how to make that work.  Would you know how to do that?  If you can pass that info down to me, I'll share it with the rest of our team.